Add the override option, now the users can ignore any other access checkers
This means, for instance if our custom perm gives access to this path:

  admin/config/media/file-system

But the user doesn't have the `administer site configuration` permission
then the users will get a forbidden error because even that the module gives
access Drupal has other access checker which denied the access.

Selecting the 'override' option any other access checker will be removed but
the module's access checker.