Remove the override checkbox.
In the end we are going to use the routes to add the access checker not the routes.

I didn't find a way to remove/add access checkers at path level just in the route subscriber.
So no need to the override option because this was anyway altering the route which could be a security issue.